Blind attempt at implementing threading for OSX
Since it doesn't support POSIX semaphores properly, we have to use mach
semaphores. Instead of sysconf to retrieve the number of CPUs, we have
to use sysctlbyname. We also have to cast the result of pthread_self
because it's a pointer.

Since mach semaphores don't support getting the current value, Windows
semaphores barely do and the feature is only used for an assertion, I
removed the function and the assertion that used it.

I have no way of testing this, so it might not work at all, it's just
for cherry picking.