Allow playback by pseudo-time
Rather than by individual messages or undo points. The paint engine is
given an amount of milliseconds to play back and it will guess how many
messages that is by assigning them weights that feel reasonable to me.
Don't have actual timing information outside of intervals, so it's not
accurate in the slightest, but it makes playback move at a much more
reasonable speed, with the ability to turn it up way faster than before.

This removes the need for passing an interval out to the playback, since
that's keeping track of its own timing properly now.